optimize engine import
