+++ packages/editor/src/functions/EditorControlFunctions.test.tsx
@@ -29,7 +29,7 @@ import { Cache, Color, MathUtils } from 'three'
 
 import { UserID } from '@etherealengine/common/src/schema.type.module'
 import { getComponent, UUIDComponent } from '@etherealengine/ecs'
-import { destroyEngine, Engine } from '@etherealengine/ecs/src/Engine'
+import { destroyEngine, Engine, startEngine } from '@etherealengine/ecs/src/Engine'
 import { EntityUUID } from '@etherealengine/ecs/src/Entity'
 import { GLTFSnapshotState, GLTFSourceState } from '@etherealengine/engine/src/gltf/GLTFState'
 import { SourceComponent } from '@etherealengine/engine/src/scene/components/SourceComponent'
@@ -37,7 +37,6 @@ import { SplineComponent } from '@etherealengine/engine/src/scene/components/Spl
 import { applyIncomingActions, getMutableState, getState } from '@etherealengine/hyperflux'
 import { HemisphereLightComponent, TransformComponent } from '@etherealengine/spatial'
 import { EngineState } from '@etherealengine/spatial/src/EngineState'
-import { createEngine } from '@etherealengine/spatial/src/initializeEngine'
 import { Physics } from '@etherealengine/spatial/src/physics/classes/Physics'
 import { PhysicsState } from '@etherealengine/spatial/src/physics/state/PhysicsState'
 import { VisibleComponent } from '@etherealengine/spatial/src/renderer/components/VisibleComponent'
@@ -47,7 +46,7 @@ import { EditorControlFunctions } from './EditorControlFunctions'
 
 describe('EditorControlFunctions', () => {
   beforeEach(async () => {
-    createEngine()
+    startEngine()
     getMutableState(PhysicsState).physicsWorld.set({} as any)
     getMutableState(EngineState).isEditing.set(true)
     getMutableState(EngineState).isEditor.set(true)
