+++ packages/engine/tests/SoA.test.ts
@@ -26,15 +26,14 @@ Ethereal Engine. All Rights Reserved.
 import assert from 'assert'
 
 import { getComponent, setComponent } from '@etherealengine/ecs/src/ComponentFunctions'
-import { destroyEngine } from '@etherealengine/ecs/src/Engine'
+import { destroyEngine, startEngine } from '@etherealengine/ecs/src/Engine'
 import { createEntity } from '@etherealengine/ecs/src/EntityFunctions'
 import { proxifyQuaternion, proxifyVector3 } from '@etherealengine/spatial/src/common/proxies/createThreejsProxy'
-import { createEngine } from '@etherealengine/spatial/src/initializeEngine'
 import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
 
 describe('Structure of Array Synchronization', () => {
   beforeEach(() => {
-    createEngine()
+    startEngine()
   })
 
   it('should synchronize values between transform objects and SoA data', () => {
