+++ packages/network/src/functions/NetworkPeerFunctions.test.tsx
@@ -28,9 +28,9 @@ import assert from 'assert'
 import { NetworkId } from '@etherealengine/common/src/interfaces/NetworkId'
 import { InstanceID, UserID } from '@etherealengine/common/src/schema.type.module'
 import { EntityUUID, UUIDComponent, getComponent } from '@etherealengine/ecs'
-import { Engine, destroyEngine } from '@etherealengine/ecs/src/Engine'
+import { Engine, destroyEngine, startEngine } from '@etherealengine/ecs/src/Engine'
 import { PeerID, applyIncomingActions, dispatchAction, getMutableState } from '@etherealengine/hyperflux'
-import { createEngine, initializeSpatialEngine } from '@etherealengine/spatial/src/initializeEngine'
+import { initializeSpatialEngine } from '@etherealengine/spatial/src/initializeEngine'
 
 import { SpawnObjectActions } from '../../../spatial/src/transform/SpawnObjectActions'
 import { createMockNetwork } from '../../tests/createMockNetwork'
@@ -41,7 +41,7 @@ import { NetworkPeerFunctions } from './NetworkPeerFunctions'
 
 describe('NetworkPeerFunctions', () => {
   beforeEach(() => {
-    createEngine()
+    startEngine()
     createMockNetwork()
     Engine.instance.store.defaultDispatchDelay = () => 0
     initializeSpatialEngine()
