+++ packages/spatial/src/input/components/InputComponent.test.tsx
@@ -34,13 +34,14 @@ import {
 import { Engine, destroyEngine } from '@etherealengine/ecs/src/Engine'
 import { ReactorReconciler } from '@etherealengine/hyperflux'
 
-import { createEngine, initializeSpatialEngine } from '../../initializeEngine'
+import { startEngine } from '@etherealengine/ecs/src/Engine'
+import { initializeSpatialEngine } from '../../initializeEngine'
 import { HighlightComponent } from '../../renderer/components/HighlightComponent'
 import { InputComponent } from './InputComponent'
 
 describe('InputComponent', () => {
   beforeEach(() => {
-    createEngine()
+    startEngine()
     initializeSpatialEngine()
   })
