+++ packages/spatial/src/renderer/components/LineSegmentComponent.test.tsx
@@ -34,7 +34,7 @@ import { destroyEngine } from '@etherealengine/ecs/src/Engine'
 import { createEntity, removeEntity } from '@etherealengine/ecs/src/EntityFunctions'
 import { getState } from '@etherealengine/hyperflux'
 
-import { createEngine } from '../../initializeEngine'
+import { startEngine } from '@etherealengine/ecs/src/Engine'
 import { ResourceState } from '../../resources/ResourceState'
 import { ObjectLayerMasks, ObjectLayers } from '../constants/ObjectLayers'
 import { GroupComponent } from './GroupComponent'
@@ -43,7 +43,7 @@ import { ObjectLayerComponents, ObjectLayerMaskComponent } from './ObjectLayerCo
 
 describe('LineSegmentComponent', () => {
   beforeEach(async () => {
-    createEngine()
+    startEngine()
   })
 
   afterEach(() => {
